+++ NEWS.md
@@ -8,7 +8,7 @@ title: "NEWS for the emmeans package"
     means that have the share letter are significantly *different*.
     We also added a vignette on "Re-engineering CLDs".
   * Bug fix for subtle error in `emtrends()` (#133)
-  * Improved customization of `emmip()` so that `we can specify`color`,
+  * Improved customization of `emmip()` so that we can specify `color`,
     `linetype`, and `symbol` are all associated with groupings; and addition of
     an example to produce a black-and-white plot. Note: While the default appearance
     of plots is unchanged, plots from your existing code may be altered if